Rework memory layout and patches
The frame is now split into patches (like before), and these patches
are rendered in columns of blocks of 16x4 pixels, from bottom to top,
and columns from left to right. Inside these 16x4 blocks, blocks of 4x2
pixels are rendered (in parallel using SIMD) in a Z-order (or actually,
more like an M-order).

The memory layout of the frame is no longer a Z-order of the pixels. The
frame is stored in memory as a regular bitmap. The frame is now stored
as RGBA instead of RGB. This wastes a quarter of the memory, but it
means that 16 pixels fit exactly into one cache line. The 16x4 blocks
that are rendered can then store four entire cache lines directly into
the backbuffer. Because the entire cache line is overwritten, there is
no need to fetch it from memory. Also, a stitching step is no longer
required: the backbuffer has the correct representation already.

There is a bug in this code: at the moment it only renders a few red
pixels to the center of the screen.