Fix BVH traversal optimisation
The issue was that I was using an "any" test whereas an "all" test
should be used. This fixes the artifacts.

The speedup over the version without this opitimization is 50 to 100%.
With this optimisation the frame rate fluctuates more heavily, because
depending on the direction of the rays, early out might save a lot, or
it might not save anything at all.