Use surface area heuristic to build the BVH
This resolves a few bugs and edge cases in my implementation. The const
parameters for the heuristic have not been tweaked, and this is all very
rough. Still, fps increased from ~45 to ~57. There is a vsync cap at 60
fps, but I am not yet hitting that. Time for some bigger meshes!