Use proper vandps and vorps instructions
It turns out that Rust does have these, but they are called `simd_and`
and `simd_or`. This appears to improve performance by a few fps to ~37
fps at this point.

The microbenchmarks show a significant improvement: from ~300 ns to ~160
ns for 8x8 intersections. Yay! Unfortunately the real win is not so big.