Fit a polynomial instead of using a Taylor series
This greatly improves accuracy on the domain (-pi, pi), at the expense
of a loss in accuracy around 0, but this is an acceptable trade off. A
Taylor series is very good at approximating a function around a point,
but when you get too far away, it breaks down quickly. Fitting a
polynomial through a few points can yield a good approximation on an
arbitrary domain, at the expense of a worse fit is the domain is too
large.