Kill some dead code, mostly related to shadow rays
I will not have shadow rays in my path tracer. It is possible to still
use shadow rays in a path tracer when doing direct light sampling, but
I don't like doing that. If you put a sheet of glass between a light
source and the surface, then everything breaks down and you get lots of
noise. I prefer a uniform treatment.