Add cosine approximation
The sixth degree polynomial might be a bit overkill for my needs, but
the fourth degree polynomial is quite inaccurate. There is also the
option of reusing the sine, but that would not allow sharing the
computations of the powers of x, which might make a combined sin and cos
faster.