Improve vector benchmarks
Put black boxes around loads and use more iterations. Also unroll the
inner loop to avoid measuring the looping overhead. The new results are
in line with the previous results, so the previous results were not
entirely wrong. (They were not memory-bound, as was the case with some
of the other benchmarks.) Still, this does improve accuracy.