Refactor multiple importance sampling code
This greatly simplifies generating rays and computing the contribution.

The resampled importance sampling when direct sampling is removed. I do
not know how to compute the probability density for direct sampling with
resampled importance sampling. What does work however, is not taking a
random triangle to sample uniformly, but weigh the samples proportional
to dot(normal, direction to barycenter).