Fix geometry factor in the BRDF
So it turns out that in all BRDF literature, there is a denominator
which is cancelled by the geometry factor. The geometry factor is never
specified, the geometry term over this factor is specified. So actually
that factor is useless, but everybody just copies it from everybody
else, it seems. I haven't found a thourough proof of why it should be
there in the first place, at best a brief note saying "normalization
factor".

In any case, I think that my direct lighting is correct now.