Fix correlations in random numbers mod 8
Because I now have 8 triangles eligible for direct sampling, my random
number generator broke down for generating indices, because the sequence
modulo 8 had only a few possibilities. In the end the resolution was
really simple: to sample an integer mod 8, take the high order bits, not
the low order bits.