Make MIS probability proportional to normal
For surfaces that are likely to benefit from direct sampling because
they face the light source, favor direct light sampling. For surfaces
that will not benefit from direct sampling because they point away from
the light source, favor BRDF sampling. This greatly reduces the number
of black pixels on surfaces that do not face the light source.