Make window opaque on Wayland
It was my own fault after al -- Wayland (or GDM, actually) just does the
bare minimum for compositing, so if I render a texture with an alpha
channel, then the window will be opaque. Simple as that. It does not
fill the background with black, that would be extra work. The solution
is to ensure that my final render is opaque, the alpha channel must be
1.0.