Pin the toolchain
I somehow missed that this had been added to Rustup, but it is a great
addition. Hopefully I will finally be able to fix the build! This
machine does not support AVX2 though, so the build fails because
build.rs happens to use the same compiler flags (ugh, Cargo should
really learn to integrate with a decent build system). I'll have to see
if it works on CI or try a different machine.